requirements.txt
ffmpeg-python==0.2.0
torchvision==0.12.0
ftfy==6.1.1
regex==2022.4.24
tqdm==4.64.0
scipy==1.8.0
pillow==9.1.0
